backend/requirements.txt
# Core Framework & ASGI Server
fastapi==0.95.1
uvicorn[standard]==0.21.1

# Database & ORM
sqlalchemy==1.4.46
asyncpg==0.27.0
alembic==1.10.2

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Task Queue & Scheduling
celery==5.2.7
redis==4.5.1

# HTTP Requests
httpx==0.23.3

# Testing
pytest==7.2.2
pytest-asyncio==0.20.3
pytest-mock==3.10.0

# Logging & Monitoring
loguru==0.7.0

# 🔹 Development Tools (Static Analysis, Formatting, Linting)
black==23.1.0
isort==5.12.0
ruff==0.1.0
mypy==1.2.0
